Back Finished

+++ LsiTask/DB/Models/BaseModel.cs
@@ -2,6 +2,6 @@
 {
     public abstract class BaseModel
     {
-        public long ID { get; set; }
+        public int ID { get; set; }
     }
 }
